fix schema error

+++ src/test/resources/long_property/schema.groovy
@@ -2,7 +2,7 @@
 schema.propertyKey("name").asText().ifNotExist().create();
 schema.propertyKey("age").asInt().ifNotExist().create();
 schema.propertyKey("city").asText().ifNotExist().create();
-schema.propertyKey("weight").asDouble().ifNotExist().create();
+schema.propertyKey("weight").asLong().ifNotExist().create();
 schema.propertyKey("date").asText().ifNotExist().create();
 
 schema.vertexLabel("person").properties("name", "age", "city", "weight").primaryKeys("name").ifNotExist().create();
